updated dir name for maze tests + updated test_cell.py
