+++ tests/maze/test_maze.py
